Added some docs for the CLI and decompose. Updated StackSVD to use less memory (cast to float32). Options to motion mode and chunksize.

+++ wfield/utils.py
@@ -198,7 +198,7 @@ class SVDStack(object):
                         self.U_warped[:,-1,:] = 1e-10
                         self.U_warped = np.stack(runpar(im_apply_transform,
                                                         self.U_warped.transpose([2,0,1]),
-                                                        M = self.M)).transpose([1,2,0])
+                                                        M = self.M)).transpose([1,2,0]).astype(np.float32)
             if not self.U_warped is None:
                 self.U = self.U_warped
                 self.warped = True
